requirements.txt
pandas==1.2.4
python-dotenv==1.0.1
requests==2.25.1
SQLAlchemy==1.4.7
psycopg2-binary==2.9.2
streamlit==1.40.1
bokeh==2.4.3
matplotlib==3.7.5
seaborn==0.13.2